fix(compiler): remove memory planning compatibility shims
