+++ packages/compiler/src/semantic/memoryState.ts
@@ -1,6 +1,7 @@
 import type {
 	CompilationContext,
 	DataStructure,
+	MemoryDefaults,
 	MemoryLayoutPlan,
 	MemoryMap,
 	MemoryPointerMetadata,
@@ -38,6 +39,11 @@ export function getPlannedMemoryDeclaration(
 	return module?.memory[memoryId];
 }
 
+interface ResolvedMemoryOverlay {
+	memoryDefaults: MemoryDefaults;
+	pointerMetadata: Record<string, MemoryPointerMetadata>;
+}
+
 function getPointerMetadataWithDefinedFields(metadata: MemoryPointerMetadata | undefined): MemoryPointerMetadata {
 	return {
 		...(metadata?.pointeeMemoryIndex !== undefined ? { pointeeMemoryIndex: metadata.pointeeMemoryIndex } : {}),
@@ -47,23 +53,31 @@ function getPointerMetadataWithDefinedFields(metadata: MemoryPointerMetadata | u
 }
 
 /** Materializes the resolved declaration shape used by codegen and public compiler output. */
-export function createMemoryItem(
-	declaration: PlannedMemoryDeclaration,
-	context: Pick<CompilationContext, 'memoryDefaults' | 'pointerMetadata'>
-): DataStructure {
-	const memoryDefault = context.memoryDefaults[declaration.id];
-	const pointerMetadata = getPointerMetadataWithDefinedFields(context.pointerMetadata[declaration.id]);
+export function createMemoryItem(declaration: PlannedMemoryDeclaration, overlay: ResolvedMemoryOverlay): DataStructure {
+	const memoryDefault = overlay.memoryDefaults[declaration.id]!;
+	const pointerMetadata = getPointerMetadataWithDefinedFields(overlay.pointerMetadata[declaration.id]);
 
 	return {
 		...declaration,
-		default: memoryDefault?.value ?? 0,
-		...(memoryDefault ? { hasExplicitDefault: memoryDefault.hasExplicitDefault === true } : {}),
-		isInherited: memoryDefault?.isInherited ?? false,
+		default: memoryDefault.value,
+		hasExplicitDefault: memoryDefault.hasExplicitDefault === true,
+		isInherited: memoryDefault.isInherited,
 		...pointerMetadata,
 	};
 }
 
-/** Returns a resolved memory item for the active module, or for a target module plan without semantic overlays. */
+function getResolvedMemoryOverlay(
+	context: CompilationContext,
+	moduleId = context.namespace.moduleName
+): ResolvedMemoryOverlay {
+	if (!moduleId || moduleId === context.namespace.moduleName) {
+		return context;
+	}
+
+	return context.namespace.namespaces[moduleId];
+}
+
+/** Returns a resolved memory item for the active module, or for a target module namespace. */
 export function getMemoryItem(
 	context: CompilationContext,
 	memoryId: string,
@@ -74,19 +88,12 @@ export function getMemoryItem(
 		return undefined;
 	}
 
-	if (!moduleId || moduleId === context.namespace.moduleName) {
-		return createMemoryItem(declaration, context);
-	}
-
-	return createMemoryItem(declaration, { memoryDefaults: {}, pointerMetadata: {} });
+	return createMemoryItem(declaration, getResolvedMemoryOverlay(context, moduleId));
 }
 
 /** Materializes a module's resolved memory map for compiler output and memory initialization. */
-export function createMemoryMapFromPlan(
-	module: PlannedMemoryModule,
-	context: Pick<CompilationContext, 'memoryDefaults' | 'pointerMetadata'>
-): MemoryMap {
+export function createMemoryMapFromPlan(module: PlannedMemoryModule, overlay: ResolvedMemoryOverlay): MemoryMap {
 	return Object.fromEntries(
-		Object.entries(module.memory).map(([id, declaration]) => [id, createMemoryItem(declaration, context)])
+		Object.entries(module.memory).map(([id, declaration]) => [id, createMemoryItem(declaration, overlay)])
 	) as MemoryMap;
 }
